*: add searcher and parallel runner
From some micro-benchmarks this is _so much faster_ than the old way of
running it, and it appears there's an additional increase in speed not
caused by just splitting the problem.

I've also added a new metric (linear) that computes the area between the
linear and non-linear terms (only when the linear term is larger) so
that the "time" spent underneath the linear component can be quantified.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>